7.0 Long IDs

+++ src/main/java/by/kobyzau/tg/bot/pbot/bots/game/exclude/ExcludeFinalizer.java
@@ -56,7 +56,7 @@ public class ExcludeFinalizer {
       return;
     }
     botActionCollector.wait(chatId, ChatAction.TYPING);
-    Set<Integer> playedIds =
+    Set<Long> playedIds =
         gameService.getExcludeGameUserValues(chatId, now).stream()
             .map(ExcludeGameUserValue::getPlayerTgId)
             .collect(Collectors.toSet());
