+++ src/main/java/by/kobyzau/tg/bot/pbot/config/dev/DevApplicationStartupRunner.java
@@ -34,7 +34,7 @@ public class DevApplicationStartupRunner implements ApplicationRunner {
   private String version;
 
   @Value("${app.admin.userId}")
-  private int adminUserId;
+  private long adminUserId;
 
   @Autowired private Logger logger;
