+++ src/main/java/by/kobyzau/tg/bot/pbot/model/ExcludeGameUserValue.java
@@ -11,7 +11,7 @@ public class ExcludeGameUserValue {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private long id;
 
-  private int playerTgId;
+  private long playerTgId;
   private long chatId;
   private LocalDate localDate;
 
@@ -23,11 +23,11 @@ public class ExcludeGameUserValue {
     this.id = id;
   }
 
-  public int getPlayerTgId() {
+  public long getPlayerTgId() {
     return playerTgId;
   }
 
-  public void setPlayerTgId(int playerTgId) {
+  public void setPlayerTgId(long playerTgId) {
     this.playerTgId = playerTgId;
   }
